Rename methods variants that return observables. Fixes #37
When observable returning variants have the same names as non-rx
variants, the compiler will require rx to be in the classpath even
though it’s a provided dependency. The compiler needs to check all
variants during compilation including the rx ones.

This commit renames observable return variants to the form
`actionUsingObservable` (ex: `getUsingObservable)` so that the methods
aren’t overloaded and rx can be used with `provided` scope as intended.